Refactor to get deepires working as a Python CLI app project * Created pyproject.toml with cleaned-up dependencies and hatchling as the build system * Removed environment.yml since deepires cannot be installed from it at least on Linux * Upgraded tensorflow to 2.15.1 and tf-keras to 2.15.0 * Using dnaio instead of BioPython SeqIO for parsing input file * Instead of reading in all sequences into memory, model prediction is run on one sequence at a time * Replaced print statements with logging * typer and rich instead of argparse for CLI arg/opt parsing and rich tracebacks if errors are encountered * Added version in __about__.py * Added pytest tests for CLI * removed deeplift dependency since it wasn't actually being used * model weights cannot be converted to Keras V3 format so sticking with Keras V2 and TF checkpoint weights
