+++ data/core.fa
