+++ data/main_independent.fa
