--- environment.yml
@@ -1,81 +1,0 @@
-name: DeepIRES
-channels:
-  - defaults
-dependencies:
-  - ca-certificates=2024.3.11=haa95532_0
-  - certifi=2022.12.7=py37haa95532_0
-  - openssl=1.1.1w=h2bbff1b_0
-  - pip=22.3.1=py37haa95532_0
-  - python=3.7.16=h6244533_0
-  - setuptools=65.6.3=py37haa95532_0
-  - sqlite=3.41.2=h2bbff1b_0
-  - vc=14.2=h21ff451_1
-  - vs2015_runtime=14.27.29016=h5e58377_2
-  - wheel=0.38.4=py37haa95532_0
-  - wincertstore=0.2=py37haa95532_2
-  - pip:
-      - absl-py==2.1.0
-      - astunparse==1.6.3
-      - bio==1.6.2
-      - biopython==1.81
-      - biothings-client==0.3.1
-      - cachetools==5.3.3
-      - charset-normalizer==3.3.2
-      - colorama==0.4.6
-      - cycler==0.11.0
-      - deeplift==0.6.13.0
-      - flatbuffers==24.3.25
-      - fonttools==4.38.0
-      - gast==0.4.0
-      - google-auth==2.29.0
-      - google-auth-oauthlib==0.4.6
-      - google-pasta==0.2.0
-      - gprofiler-official==1.0.0
-      - grpcio==1.62.2
-      - h5py==3.8.0
-      - idna==3.7
-      - importlib-metadata==6.7.0
-      - joblib==1.3.2
-      - keras==2.11.0
-      - kiwisolver==1.4.5
-      - libclang==18.1.1
-      - markdown==3.4.4
-      - markupsafe==2.1.5
-      - matplotlib==3.5.3
-      - mygene==3.2.2
-      - numpy==1.21.6
-      - oauthlib==3.2.2
-      - opt-einsum==3.3.0
-      - packaging==24.0
-      - pandas==1.3.5
-      - pillow==9.5.0
-      - platformdirs==4.0.0
-      - pooch==1.8.1
-      - protobuf==3.19.6
-      - pyasn1==0.5.1
-      - pyasn1-modules==0.3.0
-      - pyparsing==3.1.2
-      - python-dateutil==2.9.0.post0
-      - pytz==2024.1
-      - requests==2.31.0
-      - requests-oauthlib==2.0.0
-      - rsa==4.9
-      - scikit-learn==1.0.2
-      - scipy==1.7.3
-      - six==1.16.0
-      - tensorboard==2.11.2
-      - tensorboard-data-server==0.6.1
-      - tensorboard-plugin-wit==1.8.1
-      - tensorflow==2.11.0
-      - tensorflow-estimator==2.11.0
-      - tensorflow-intel==2.11.0
-      - tensorflow-io-gcs-filesystem==0.31.0
-      - termcolor==2.3.0
-      - threadpoolctl==3.1.0
-      - tqdm==4.66.2
-      - typing-extensions==4.7.1
-      - urllib3==2.0.7
-      - werkzeug==2.2.3
-      - wrapt==1.16.0
-      - zipp==3.15.0
-prefix: C:\Users\czw20\.conda\envs\DeepIRES
