+++ src/deepires/model/__init__.py
@@ -1,0 +1,2 @@
+from deepires.model.model import deepires_model
+from deepires.model.sequence_encode import get_data_onehot
